ok, the names are bad, but (in italics) are they
