+++ client/components/Editor/commands/index.ts
@@ -3,3 +3,4 @@ export * from './marks';
 export * from './rtl';
 export * from './tables';
 export * from './textAlign';
+export * from './splitBlock';
