+++ client/components/Editor/plugins/keymap.ts
@@ -18,7 +18,7 @@ import { undo, redo } from 'prosemirror-history';
 import { undoInputRule } from 'prosemirror-inputrules';
 import { keymap } from 'prosemirror-keymap';
 
-import { splitBlockPreservingTextAlign } from '../commands';
+import { splitBlockPreservingAttrs } from '../commands';
 
 const mac = typeof navigator !== 'undefined' ? /Mac/.test(navigator.platform) : false;
 
@@ -142,7 +142,7 @@ export default (schema) => {
 		newlineInCode,
 		createParagraphNear,
 		liftEmptyBlock,
-		splitBlockPreservingTextAlign,
+		splitBlockPreservingAttrs(['textAlign', 'rtl']),
 	);
 
 	return [keymap(keys), keymap({ ...baseKeymap, Enter: customEnterCommand })];
